[Refactor] move out the os-related string utility into `smartStrings.ts`
